pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sraosha"]

[project]
name = "sraosha"
version = "0.1.0"
description = "Enforcement and governance runtime for data contracts"
requires-python = ">=3.11"
license = { text = "MIT" }
readme = "README.md"
authors = [{ name = "Sraosha Contributors" }]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "duckdb>=0.10.0",
    "networkx>=3.3",
    "celery[redis]>=5.3.0",
    "pydantic-settings>=2.2.0",
    "typer[all]>=0.12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "numpy>=2.1.0",
]

[project.scripts]
sraosha = "sraosha.cli.main:app"

[project.optional-dependencies]
datacontract = ["datacontract-cli>=0.10.0"]
airflow = ["apache-airflow>=2.7.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "aiosqlite>=0.20.0",
    "pre-commit>=3.7",
]
all = ["sraosha[datacontract,airflow,dev]"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
